Add files via upload
Wine Quality Prediction – Machine Learning
Here we will predict the quality of wine on the basis of giving features. We use the wine quality dataset . This dataset has the fundamental features which are responsible for affecting the quality of the wine. By the use of several Machine learning models, we will predict the quality of the wine. Here we will only deal with the white type wine quality, we use classification techniques to check further the quality of the wine i.e. is it good or bed.